[FIX] point_of_sale: prevent the new session sequencing with prefix to start with the config name
Description of the issue/feature this PR addresses:
The sequencing for new PoS Session isn't right when set to another value than the default.

Current behavior before PR:
The PoS name was added before the intended sequence.
For example, if you set the pos.session sequence to TEST/ with the Furniture Shop, it return Furniture ShopTEST/00001.

Desired behavior after PR is merged:
The PoS Session name should be using only the sequence if the sequence isn't the default one.
When the pos.session sequence is set to TEST/ and you open the Furniture Shop, it should return TEST/00001.

opw-4822673