[IMP] website: display effective write uid and date for website pages
This is a follow-up of [1].

The "page" list view (`website.page` model) has the possibility to
display the `write_uid` and `write_date` fields. It was misleading as it
was not about the last modification of the page content, but the last
modification of the page record itself (e.g. when the page URL is
changed). This commit creates two new fields to compute the "effective"
write uid and write date, which are about the last modification of the
page record *or content* (inner `ir.ui.view`'s arch).

Note: it was discussed with the framework-py team that this should not
be an optional feature of the `inherits` mechanism, but a case by case
implementation as it is done here.

[1]: https://github.com/odoo/odoo/commit/20a211de39a8db7422e9db83951652681820935c